BUOI 2/ex01/lab-05/video-hidden/requirements.txt
Pillow
opencv-python
cryptography